+++ src/main/java/codesquad/web/handler/HomeForwardHandler.java
@@ -1,13 +1,14 @@
-package codesquad.web;
+package codesquad.web.handler;
 
 import codesquad.was.http.HttpRequest;
 import codesquad.was.http.HttpResponse;
+import codesquad.web.RequestHandler;
 import java.io.IOException;
 
-public class MainForwardHandler implements RequestHandler {
+public class HomeForwardHandler implements RequestHandler {
 
     @Override
     public void process(final HttpRequest request, final HttpResponse response) throws IOException {
-        response.forward("/main/index.html");
+        response.forward("/index.html");
     }
 }
