+++ src/main/java/codesquad/web/handler/StaticResourceForwardHandler.java
@@ -1,7 +1,8 @@
-package codesquad.web;
+package codesquad.web.handler;
 
 import codesquad.was.http.HttpRequest;
 import codesquad.was.http.HttpResponse;
+import codesquad.web.RequestHandler;
 import java.io.IOException;
 
 public class StaticResourceForwardHandler implements RequestHandler {
