+++ src/test/java/codesquad/web/handler/SignUpRequestHandlerTest.java
@@ -1,4 +1,4 @@
-package codesquad.web;
+package codesquad.web.handler;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
